requirements.txt
fastapi
uvicorn
weaviate-client
openai
requests
numpy
mysql-connector-python
